:recycle: Reorganized files/structure
