+++ src/components/Navigation/TopBar/Help/Feedback/Feedback.const.ts
@@ -2,7 +2,7 @@ import {
   FeedbackLocalStorage,
   SlackStatus,
   StatusEnum,
-} from './FeedbackForm.types';
+} from './Feedback.types';
 
 export const MAX_FILE_SIZE_BYTES = 1048575;
