+++ src/components/Navigation/TopBar/Help/Feedback/Feedback.tsx
@@ -1,0 +1,20 @@
+import { FC } from 'react';
+
+import FeedbackContextProvider from './providers/FeedbackContextProvider';
+import { FeedbackType } from './Feedback.types';
+import FeedbackInner from './FeedbackInner';
+
+interface FeedbackProps {
+  onClose: () => void;
+  selectedType: FeedbackType;
+}
+
+const Feedback: FC<FeedbackProps> = ({ onClose, selectedType }) => {
+  return (
+    <FeedbackContextProvider onClose={onClose} selectedType={selectedType}>
+      <FeedbackInner />
+    </FeedbackContextProvider>
+  );
+};
+
+export default Feedback;
