+++ src/components/Navigation/TopBar/Help/Feedback/FeedbackInner.tsx
@@ -1,0 +1,16 @@
+import { FC } from 'react';
+
+import ResponsePage from './components/ResponsePage/ResponsePage';
+import { useFeedbackContext } from './hooks/useFeedbackContext';
+import FeedbackForm from './FeedbackForm';
+
+interface FeedbackInnerProps {}
+
+const FeedbackInner: FC<FeedbackInnerProps> = () => {
+  const { showResponsePage } = useFeedbackContext();
+
+  if (showResponsePage) return <ResponsePage />;
+  return <FeedbackForm />;
+};
+
+export default FeedbackInner;
