+++ src/components/Navigation/TopBar/Help/Feedback/hooks/useServiceNowIncident.ts
@@ -1,0 +1,13 @@
+import { useMutation } from '@tanstack/react-query';
+
+import { PortalService } from '../../../../../../api/services/PortalService';
+import { SERVICE_NOW_QUERY_KEY } from '../Feedback.const';
+import { FeedbackContentType } from '../Feedback.types';
+
+export function useServiceNowIncident(feedbackContent: FeedbackContentType) {
+  return useMutation({
+    mutationKey: [SERVICE_NOW_QUERY_KEY, feedbackContent],
+    mutationFn: async (formData: FormData) =>
+      PortalService.createIncident(formData),
+  });
+}
